Add events to track webchat interaction
It would be good for us to be be able to track how users are interacting
with web chat. This would allow us to be able to pull paths that users
went through that ended up with an accepted web chat. I have chosen to
track errors from the egain script as well so that we can show how many
missed offers were problems with communciating with KCOM or eGain.

I have standardised the messaging backwards and forwards between frames
as well so that they both pass stringified objects.