Remove sandboxing of the webchat iframe
After trying to use the chat it seemed like we would need to give the
iframe all of the permissions that sandboxing allowed, so it seemed like
we would not gain anything from using it. As we do trust this 3rd party
a little more than we would an ad network this is probably fine, they
still can't play with our DOM and such.