Revert back to a normal script loaded
The jQuery method of loading scripts doesn’t create a script element in the body. The egain code was calling back to the body to see the script that loaded it so that it could see where it was being called from, hence was failing.

As IE doesn’t understand the load event on scripts we can use the readystatechange event as well to bring it into line.