Fix HMRC webchat offer
The iframe was relying on finding an element with a specific id in the offer markup loaded from KCOM. At some point recently they have updated their markup with a new id, causing the offer to fail to load.

This changes the code such that it now just looks for the first div inside the body rather than relying on a specific id which appears likely to change in future.