Add aria roles and labelledby to related items
When passing in a section id, use this to add an id attribute and a
labelledby attribute. This is in favour of generating an id, eg. by
parameterising the title so that the component isn't using rails specific
features.
Care must be taken when using the component so that navigation is always
correctly labelled.
Pattern based on http://tink.uk/screen-readers-aria-roles-html5-support/