Allow text-only breadcumbs
User testing conducted by the Service Manual team has shown that users
understand breadcrumbs better when the last breadcrumbs element references
the current page, thus giving the user a mental anchor.

Allow passing in a breadcrumb hash with no URL that renders it as a text
element. Only the last breadcrumb element should be text-only. There are not
checks in code to enforce it, so it's left to a developer of each application.