Add WH attachment styling to Govspeak component
As part of migration we are moving formats from WH to `content-store`
and `government-frontend`. Some of the formats in WH have `body`
content that includes embedded attachments, eg, markup describing
an attachment in the rendered HTML of a document from content-store.

To render this in a new-world frontend app, like government-frontend
the `govspeak` component, that renders a documents `body` content,
needs to be able to add styling and behaviour for the attachment
markup in the `body`.

This PR adds examples of attachments within govspeak, and ports the
CSS/JS that renders them/adds toggling behaviour, to the govspeak
component on static.

- Includes changes to work with accessible JS toggle module [1]
- Fixup the attachment SCSS to pass the new linting rules, mostly
  whitespace and not-referencing-colours-by-variable-name.
- Switch the non-standard 1/6 margin (only used in WH, not FET)
  to something specific to it's use here, avoid introducing the
  additional margin size to static.

[1] https://github.com/alphagov/whitehall/pull/2476

Note, the CSS to styling attachment govspeak is still duplicated in
WH and at some point will need to be reconsiled with the component
version, but that needs to be done for all govspeak styling, not
just attachments. But for now we're adding the ability for the
component version to render attachments at all, to unblock the
migration of formats.

Also note, the static images used to represent attachments without
their own cover image, eg doc, csv, html pub, have full URLs which
point at the WH-frontend app's assets. This is the approach we're
taking short term, long term the image assets should belong to
something within the FE stack (most likely static). The approach of
storing fully rendered attachment HTML in the content store is a
interim approach to support migration, and will be replaced with
a system where the data for an attachment is passed to a frontend,
with the body content containing a placeholder, and something in
the FE stack owning the template to render the attachment, and
inserting the output into the body for rendering. Keeping the
responsibility/ownership of attachment template and images within
the FE stack, not partially owned by Whitehall.