Don't require class selector for JS toggle module
We can detect the control/toggle element from the module data-*
attributes, and don't need to rely on a specific class as well.

This feels more in keeping with the data-module approach we're taking
where the 'interface' to the module is encoded in data-attributes
and their values, and doesn't rely on anything else in the element
passed into the module, eg classes.

This allows the class of the element to be specific to use case in
the app invoking the module.

I've removed the class part of the selector, and updated the tests
so the examples use an example selector that is less likely to be
confused for module configuration.