Use `js-hidden` in tests, to match app usage
A conditional hook to hide content if JS is enabled is being added
to `govuk_template`, and is a pattern we use a lot in other apps.

Doing this, rather than hiding the content with the module JS
prevents a flash of visible content between the content loading and
the JS executing.