Add examples of Govspeak fractions to component
The styling already existed, having been ported from Whitehall
when the Govspeak component was first ceated.

The migration of HTML Publications to `government-frontend` means
the fractions SCSS in the component will be used in production and
we should have some examples of how they render, to illustrate the
pattern and make testing regressions easier.