Fine tune to match existing Whitehall rendering
* Coerce logos to have matching heights. The crest images are all
slightly different sizes. Allow the width to vary, but alter the
background size so that they are either 25px or 35px, as per Whitehall.
This eliminates the need for the more complex crest mixin that needs
background sizes.

* Tweak the mobile padding to match Whitehall - crests are 20px tall,
expanding to 25px on tablets, tall crests start at 25px and expand to
35px.