Fix IE8 JavaScript error on GOV.UK
IE8 is throwing an error on all of GOV.UK (reported by @maxf). Looking at the given line/column number reveals a use of indexOf in the webchat code. This isn’t supported by IE8, hence it’s throwing an error. We’ve got jQuery available, so the fix is to swap in $.inArray().