Add sticky-element-container module.
This module allows you to stick an element to the bottom of its parent.

This is part of migrating `html_publications` from `whitehall` to
`government-frontend`, and is meant to be used to implement the little
"&uarr; Contents" button on
http://www.gov.uk/government/publications/dwp-single-departmental-plan-2015-to-2020/dwp-single-departmental-plan-2015-to-2020.

Another PR will be opened to `government-frontend` to update the content
item template and pass in the correct parameter to use this. (Update:
https://github.com/alphagov/government-frontend/issues/122)

Relevant Trello ticket:
https://trello.com/c/cJGrtnc1/286-5-html-publications-migration-front-end-work-large

Previous implementation in Whitehall:
https://github.com/alphagov/whitehall/blob/master/app/assets/javascripts/application/back-to-content.js

The module is called `sticky-element-container` in order to allow future
additional behaviours. Most notably, the ability to have multiple sticky
children fixed to different parts of the screen.

BEM is used for the style modifiers to avoid polluting the global CSS namespace
with potentially clashing stateful classes such as `hidden`.

The component allows overriding of two functions which query the window.
These are used to aid with stubbing in the unit tests.