Update emergency banner link styling.
The link will overlap the text if the last line of the text is too long.

This causes the link to display as it does on mobile, which is left
aligned and block-level. It also changes to a semantic CSS class name.

Relevant Trello ticket:
https://trello.com/c/cm6VO0q4/341-emergency-publishing-frontend-bugs-medium

A follow-up PR will be submitted to alphagov/fabric-scripts to update
the class name.