Include inline script for toggling bar class
To prevent the global bar from flashing or shifting the page up and
down, we want to check whether to show the bar as soon as possible, and
set a class if we want to show it.

* Include the full source in the test
* Run the test spec against source and the minified version

Ideally we would have the full source minified on the fly and injected
into the template.