Add register-to-vote promo on transaction done pages
Only on completed transaction format (though we'll extend it to transaction
pages in the near future). Exclude register to vote done page, as there is
not a lot of point promoing something you've just done.

As some done pages will have register-to-vote links in their related links
we need to exclude those, and if they're the only link in a group, the
group as well.

Filtering the register-to-vote link out of the related_artefacts before
grouping makes this much easier, but should only be done for pages where
we're showing the promo.

Note: register-to-vote is filtered out by content id, as the related
artefacts don't have path/slug field, and the full web_url is pretty
long, and can vary per environment, so use the content_id, which should
be stable.

Note: the changes to the tests are not super elegant, but this change is
likely to be reverted wholesale once the promo is complete, this isn't a
pattern we're planning to keep around.