Do not assume the youtube API is using the same protocol
Phantom's window object has a default location of `file:///` which has a protocol of `file:`.

In the AccessibleMediaPlayer the following snippet determine's which type of player to load.

```
if (config.url.match(/^(http|https)\:\/\/www\.youtube\.com/)) {
  // a youtube-y player
} else {
  // a non youtube-y player
}
```

So if we get to that point in a jasmine/phantomjs test and our protocol is `file:` rather than `http:` or `https:` it's going to attempt to load a non-youtube-y player reagardless.