Fix search box breaking when increasing font size
When something has a height (especially of px) it doesn't grow
with the text (not when zooming but increasing just the font size).
Removing the height completely would be the ideal solution
but that caused issues in IE9, IE10 and WindowsPhone due to the alignment
with the submit button.
The `height` got substituted with `line-height` but that alone
caused issues in Safari in many iOS versions and Chrome
on some Android phones, the text box somehow had a height of 36px
and not 35px which caused a white line to appear below the submit button.
That's why the height of all the things got increased to 36px
which fixed it. (Although it is not clear what caused it.)