Fix transaction bug introduced in 8f1f835
Attempting to refactor the transaction detection logic, I inverted the
match (returning true if inside a transaction URL, not false if *not*
inside a transaction URL) but forgot to invert the if statement to "OR"
from "AND".

This commit fixes that, and adds some tests to ensure the mistake can't
happen again.