Handle path comparison better.
The intention here is to return true for the two forms of paths.

Example:
1) /register-to-vote
2) /register-to-vote/

Rightly the previous code only checked and returned true for first instance and failed in the second.

This behaviour was only observed whilst working with the dev VM.

On Integration, Staging and GOV.UK paths are very consistent and always resulting/returning the aforementioned first instance.

I have added this method to better cater for both, It splits the pathnames with the forward slash delimiter and joins all.

This in effect returns a combination of all the characters in the path for the current page, returning true for only paths that equal the defined slug.