Migrate link href printing from template
govuk_template prints relative links with a forced gov.uk prefix. This works for gov.uk but not for other services that consume the template. First step to fix this is to insert the prefix as part of static. After this is live we can remove the prefix override from the template (i.e. change the link href printing selector to a[href^="/"]:after, a[href^="http://"]:after, a[href^="https://"]:after).