Add breadcrumb meta tag
This commit updates the list of existing meta tags with one for breadcrumb,
and also defines a function for sending the breadcrumb to google analytics.

This is being implemented upon request from the performance analyst.

This will assist with the production of Alpha reports, based on topic and page
performance.