Script to copy translations from Whitehall
Components need translations for some of their labels. For now,
the best source of these is Whitehall. Not all labels are translated
but due to the way i18n works it's ok to leave an empty string for
a translation as it'll fallback to either `en` or the supplied default.