Make sure breadcrumbs and related items look good in older apps
The GOV.UK Components expect that they're included in a page with a
certain global CSS reset. This is the case for newly migrated pages
(like the ones rendered by government-frontend), but it doesn't apply
to pages in older frontend apps like calendars, calculators, smart
answers, and others.

While migrating those pages to use the component we introduced the same
CSS in all applications. This to prevent extra margins appearing on
breadcrumb and related links.

https://github.com/search?q=org%3Aalphagov+%22might+be+pushed+up+to+stat
ic+at+some+point%22&type=Code

This puts the CSS back in here, so that we can remove the overrides in
individual apps.

https://trello.com/c/5Utjs9qw