Move the education_survey to the top of the smallSurveys list
The algorithm for picking which survey to display on a page is to go
through all the surveys in smallSurveys and ask each one in turn if
it should be on this page - only taking dates and activeWhen into
account.  We choose the last one that says yes.  This means that if
several surveys could be shown we will pick the one that is last in
the list.

The education survey and the 2nd user research survey would both
trigger on the same page, but the 2nd user research survey should "win"
because it is:

a) more specific - it matches on a single url compared to the education
survey which matches on a bunch of possible things
b) more frequent - it has 1/1 frequency, compared to the education survey
which has 1/10 frequency

There's no way currently to take this into account, so we reorder the
list so that the education survey is first in the list and the more
specific / more frequent ones come later.  A future story may change
the algorithm to pick all surveys that could match and then show the
first one that would display based on frequency.