Use `attr` instead of `data` to access properties
This commit changes the way we access data attributes when sending click events
to Google Analytics.

In some cases, some data attributes can be numbers encoded as strings. It is
important to preserve the type of these attributes so Google Analytics doesn't
complain about it.

When we use `data()`, according to jQuery's documentation, this happens:

```
Every attempt is made to convert the string to a JavaScript value (this includes
booleans, numbers, objects, arrays, and null).
```

This means that when tracking breadcrumb actions, which are numbers, they are
converted from string to numbers and the Google Analytics' API returns:

```
Expected a string value for field: eventAction. But found: "number".
```