Fix breadcrumb and section matchers for Education survey
In #879 we replaced the `#global-breadcrumb` container for breadcrumbs
with the govuk component rendered version which uses `.govuk-breadcrumbs`
as the container instead.  This broke our `breadcrumbMatches()` function
so we change the function to grab the correct selector.

We also change the section matcher to use a regexp test instead of direct
equality to check for a matching section.  It's not clear if the section
names have recently changed, or if the matcher never worked but it wasn't
detected because the breadcrumbs worked instead.