Add analytics dimension for A/B tests
Link meta tags named `govuk:ab-test` to a custom Google Analytics
dimension. The value of the meta tag and the dimension is a combination
of the test name and bucket name, such as `EducationNavigation:B`.

This links to existing Session dimension 13, which has been used for A/B
tests in the past.

Note that only one A/B test can use this dimension at once, otherwise
the value will be overridden by different tests.