Add new environment short survey
Note that at the time of writing this survey is running concurrently
with another survey that affects a small number of pages. It was decided
to retain the existing survey for these pages rather than choose between
multiple candidate surveys per page. By placing the new survey at the top of the
array we ensure that it will be "overridden" by a survey that also
applies where that survey is found later in the array.

Trello: https://trello.com/c/J8PdbakJ/108-gov-uk-survey-environment-short-survey