Remove ‘information block’ feature
The information-block class was originally added when call-to-action was added to whitehall, and a specialist guide was refactored so that `call-to-action` and `information-block` shared some common CSS[1].

The last reference to the information-block within templates in whitehall was removed when 'detailed guides' were moved to government-frontend[2].

`information-block` itself does not appear to have ever been used in HTML generated by the Govspeak engines in whitehall nor in the standalone gem.

Searching the commit history for whitehall returns only mentions of “information-block” within templates, CSS and tests, and searching the history for alphagov/govspeak returns nothing.

At time of commiting, according to GitHub’s search, there are only 11 files mentioning information-block across the entire alphagov organisation. Ten of these are CSS files, the other is the govspeak.yml file in this repo.[3]

Lastly, we grepped the mirrors for mentions of information-block and found only three results:

```
./guidance/animal-welfare-in-severe-weather.html:        <aside class="related-mainstream-content information-block" role="complementary">
./guidance/disclosure-and-barring-service-criminal-record-checks-referrals-and-complaints.html:        <aside class="related-mainstream-content information-block" role="complementary">
./guidance/pet-travel-quarantine.html:        <aside class="related-mainstream-content information-block" role="complementary">
```

However these all appear to be outdated as the class is not present when viewing the corresponding webpages in production.

[1]: https://github.com/alphagov/whitehall/commit/d65866434510960f98ec4d2afa968c41b3eb9c6f
[2]: https://github.com/alphagov/whitehall/commit/08b9056fb1d19b0fd3b5f29f6f72f6a34dd263fc
[2]: https://github.com/search?q=org%3Aalphagov+information-block&type=Code