Add block of Google Analytics dimensions for A/B tests
The GA dimensions from 40-49 have been allocated to A/B tests. These may
be reserved for any A/B test for the duration of the test.

This commit identifies which dimension to use for the A/B test based on
an extra `data` attribute added to the `meta` tag. Only dimensions in
the allowed range are permitted. Any others are ignored and are not
reported to GA.