Add first email survey
This is actually 2 surveys.  First we run a 1-in-1 frequency running for from
when we deploy it (Thursday 9th) to 1pm on Friday 10th with identifier
`govuk_email_survey_t01`.  We then run a 1-in-2 survey from 1pm on Friday 10th
to the end of Monday 13th, this uses the same `activeWhen` rules but has a
different identifier.  We use different identifiers because we want them to
be distinguishable at the analytics level, and to send different urls to the
users (the urls will take the users to the same survey, but allow us to collect
them into different buckets based on the url they cam in via).  We have set up
these urls in two EmailSurvey instances in the feedback app, configured to run
from the 8th to the 10th and from the 10th to the 13th.

Once we've looked at the figures we can deploy again on Monday to run for a
longer period or different frequency or targeting different pages.

The activeWhen is interesting in that it's an exclusion rather than
inclusion filter - e.g. all pages *except* the education ones.