Update scroll tracker for 2017 budget
HMT have asked for scroll tracking to be added to the 2017 budget publication. They want to track the numbers of users who reach the links in the headings I've added to the code.

This pull request replaces #925 (I've created a new one as the headings' titles have changed and this request reflects the change).