Add documentation mentioning feedback version of email surveys
The id and dates for an email type survey are defined twice.  Once in
static to decide if we should show the UI to a user, and once in feedback
to define what url to send to the user in the email.  This commit adds
documenation to make it clear that defining an email survey in static also
needs its counterpart to be defined in feedback.