Explicitly include button styles for the email survey submit button
The send button in the email survey UI used the ".button" class to get the
styling defined in `helpers/_buttons`.  Unfortunately this styling is not
included in all layouts that an app might request from static.  For example
it is included in the static layout, but not the header_and_footer_only or
core_layout.  This means that only some frontend rendering apps will style
the button correctly.

Instead we explicitly include the button styling mixin in the css for a
button in the user-satisfaction-survey in `helpers/_header`.  This helper
is included in all layouts so all frontend rendering apps will style the
button correctly.