Move survey templates to the view and parameterise them
First we move the two templates for the email and url survey types out of
the surveys.js and add them as script tags that live inside the
user-satisfaction-survey-container div.  We use script tags so we don't
need to hide them with CSS, and it's more obvious that these are templates
not full bits of content.

We change the survey definitions to remove the `template` property and
introduce a `templateArguments` property instead.  This means surveys can
no longer completely change the UI, but can only customise the existing
UI.  For all the surveys we've so far delivered, any custom `template`
propety has only been to change the text, so we don't think this'll be a
problem.  Default values for the customisable text are included as data
attributes on the script tag that contains the template.

In the JS implementation we instantiate the templates only when we get to
the point where we want to display a survey, which should keep things
snappy when no surveys are involved.