Use strings to define startTime and endTime in surveys
Again, our goal is to be able to define the surveys outside of the JS file
so we can't set `startTime` and `endTime` to calculated values like
`new Date('...').getTime()`.  Instead we leave them set to the strings we
were passing into `new Date('...')` and parse this at the point we
actually need it.