Move survey definitions to HTML
Instead of defining Javascript objects for the surveys inside the JS
bundle, we define them in script tags in the HTML.  We use script tags
again because they don't have display concerns.  The default survey
has a `data-survey-default` attribute, and this is replaces the
`defaultSurvey` javascript object.  All other surveys are read in from
elements with a `data-survey` attribute, and these replace the
`smallSurveys` array of javascript objects.  In all other respects the
functionality is the same.  We've added a few more guard clauses to the
survey implementation to make sure that it doesn't break if there are
no surveys defined.

We've used JSON inside the script tags, and JSON.parse to interpret them.
Given that the definition of a survey can be quite complex, this felt like
an easier solution than using data attributes for everything.