Provide standard mechanism for including currentPath in survey url
The survey implementation had a special case to add `?c=<currentPath>` to
the survey url if and only if it was a url to a surveymonkey suryvey.
We've seen other surveys that needed this behaviour, but we've done it
dynamically in the definition.  As that will no longer be possible, we add
code to replace a {{currentPath}} template in the URL with the currentPath.

Note that we also update the url for the default survey, which was a
surveymonkey url, so it continues to receive the param.