Show all related links provided in the data
Note: this is now being done in govuk_navigation_helpers:
https://github.com/alphagov/govuk_navigation_helpers/pull/62

The reasons for that are described in that pull request. Basically, we
need flexibility in the way we present related links. The previous rule
here was to only display related links for the first 2 sections of the
side bar. That's the correct behaviour, but we now need to display links
for sections such as "Elsewhere on GOV.UK", which appear below the
taxonomy sections. That means in certain circumstances, we might not
be showing them because of the index of those sections.

The logic of what should presented should be provided by navigation
helpers in order to give us that flexibility.