Change how emergency banner is stored in Redis
Rather than using a single `enabled` key to check if the emergency
banner can be displayed, instead get the contents of an
`emergency_banner` hash from Redis. If the contents of the hash contains
a value for campaign class, display the banner.

As we get the content for more parts of the banner, the expectation in
`enabled?` will be updated.