Refactor the EmergencyBanner class
The code in the `EmergencyBanner` class has been moved to
the `Display` class in the `EmergencyBanner` module.

This is a preparatory step to allow functionality to be added to the
EmergencyBanner module to set the values for the emergency banner.
For example, adding values to Redis to “deploy” the banner and deleting
them to “remove” it.

All tests have been updated and are still passing.
All existing lint errors have been fixed.