Memoize the connection to Redis
Memoize the connection  to Redis so that we’re not setting up a new 
connection on every request.

Add a timeout on the Redis connection of 0.1 seconds, so that the user
doesn’t have to wait seconds for the page to load if the connection to
Regis is slow. Setting the timeout ensures that the page still loads, 
even if it’s without the emergency banner. The default `timeout` for 
Redis is 5 seconds.