Track total number of sections and section links
At the moment we only record related link and accordion link information
on the click event (ie only if someone uses the link). We have no
information about if links are not used, since they don’t get clicked
and then never send a tracking event.

This change will add tracking for custom dimensions on the page view,
which will report information on:

- the total number of “sections” on the page
- the total number of “section links” on the page

These values will have different meanings depending on the type of page
we are on.

On a content item, the number of sections will refer to the number of
sections of related links in the sidebar. The total number of links
is the total number of related links in the sidebar (not including
the heading in a taxon sidebar, but including the “More” link).

On a leaf page in the new taxon navigation, the number of sections is
not tracked, and the total number of links is the total number of
leaf links.

On an accordion page in the new taxon navigation, the number of
sections will refer to the number of accordion sections, and the total
number of links will refer to the total number of content items
under all the accordion sections.

On a grid page in the new taxon navigation, the number of sections is
the number of grid links, and the total number of links will refer to
the total number of grid links plus the leaf links.