Change how the contents from Redis are memorised
Memoize the call to get the `emergency_banner` key, so that if the key
doesn’t exist and returns a nil, this nil value will be memoized and
Regis won’t potentially be called 4 times per page view.

Using if defined? rather than ||= means that the @content variable is
checked to see if it has been defined at all, regardless of whether it's
nil or not.

Also notify Erribt if the call to Redis fails.