Track page views on DOM ready
We have a bug that means that some page views do not correctly have
the number of links and sections set. This could be because these values
are tracked through counting DOM elements, but the code is run before
`document.ready`.

This change attempts to fix the bug by performing the count after the
DOM has fully loaded, and only then tracking the page view.