Pass custom dimensions directly to page tracking
We’re seeing some weird behaviour on our Google Analytics tracking:
- Some pages appear to have a very high (10-100) repeat visit rate
  within a single (half hour) session
- Dimensions that rely on the DOM having loaded also are reporting 0
  count sometimes.

Noteably, when the count *is* reported correctly, the repeat visit
rate also appears to drop back to “normal levels” (~2).

This could be because of setting custom dimensions at the “session”
level, rather than passing them directly into the page view tracking,
but this is hard to test because we cannot reproduce the bug on our
machines.

In order to debug, we’re attempting to pass custom dimensions
directly to the page view tracking.