Enable setting options for the next tracked pageview
We currently have some generic methods for calling arbitrary methods
within the Analytics object on the next page view. This is actually only
ever used to set a custom dimension.

In order to close down our API, this change removes this generic,
reflective and overly open method, and instead exposes a method to set
an options object for the next tracked pageview, which is really what
we’re trying to achieve.

This will enable us to also remove the externally-exposed, specialised
dimension-setting methods.