Add Navigation Tracking Metatags shim
In order to remove the navigation analytics meta tags, we need to do
two stages:

1. Remove all references to the meta tags (ie in taxonomy sidebar
   and related items)
2. Remove the component itself

This is because Static/Slimmer will have a cached set of sidebar
components that will continue trying to use the old meta tags. Hence
we need to leave the meta tags in place, and then remove them after
the cache has cleared.