Extract surveys.js as a separate JS bundle
The contents of surveys.js can change frequently as we introduce new
surveys and retire old ones.  Every time we deploy these changes the
application.js bundle gets a new digest hash and so frequent visitors
have to re-download the whole JS bundle just to get these updates to
surveys.  To avoid this we separate the surveys.js so it is required
separately and changes to it will not change the digest of the main
application.js.

The assumption here is that applications using the JS provided by
static only do so by using the layouts via slimmer and not by requesting
the application.js or header-footer-only.js files directly.  If that
is the case they will also need to be updated to also request the
surveys.js bundle.

We also take the opportunity to make the initialization of the surveys
JS code more similar to other components.  We use window.GOVUK instead
of root.GOVUK and we pass in jQuery rather than extract it from root.